+++ core/src/dev/suap/breakout/interfaces/Collidable.java
@@ -1,10 +1,10 @@
 package dev.suap.breakout.interfaces;
 
 public interface Collidable {    
-    int getLeft();
-    int getRight();
-    int getBottom();
-    int getTop();
+    float getLeft();
+    float getRight();
+    float getBottom();
+    float getTop();
     
     void markAsColliding();
     void resolveCollision();
